requirements.txt
asgiref==3.7.2
Django==5.0.2
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
Markdown==3.5.2
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.4.4
